Added mechanism to interrupt kopia operation during the test run
