+++ tests/recovery/recovery_test/recovery_test.go
@@ -7,14 +7,16 @@ import (
 	"errors"
 	"log"
 	"os"
+	"os/exec"
 	"path"
 	"strings"
+	"syscall"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/require"
 
 	"github.com/kopia/kopia/tests/recovery/blobmanipulator"
-	"github.com/kopia/kopia/tests/robustness/snapmeta"
 	"github.com/kopia/kopia/tests/tools/kopiarunner"
 )
 
@@ -27,12 +29,7 @@ func TestSnapshotFix(t *testing.T) {
 		t.FailNow()
 	}
 
-	ks := getSnapshotter(baseDir, dataRepoPath)
-	if ks == nil {
-		t.FailNow()
-	}
-
-	bm, err := blobmanipulator.NewBlobManipulator(baseDir)
+	bm, err := blobmanipulator.NewBlobManipulator(baseDir, dataRepoPath)
 	if err != nil {
 		if errors.Is(err, kopiarunner.ErrExeVariableNotSet) {
 			t.Skip("Skipping recovery tests because KOPIA_EXE is not set")
@@ -40,13 +37,6 @@ func TestSnapshotFix(t *testing.T) {
 			t.Skip("Error creating Blob Manipulator:", err)
 		}
 	}
-
-	bm.DirCreater = ks
-
-	// err = bm.ConnectOrCreateRepo(dataRepoPath)
-	// if err != nil {
-	// 	t.FailNow()
-	// }
 
 	bm.DataRepoPath = dataRepoPath
 
@@ -57,6 +47,13 @@ func TestSnapshotFix(t *testing.T) {
 		t.FailNow()
 	}
 
+	kopiaExe := os.Getenv("KOPIA_EXE")
+	cmd := exec.Command(kopiaExe, "maintenance", "run", "--full", "--force", "--safety", "none")
+	time.AfterFunc(10*time.Millisecond, func() {
+		syscall.Kill(cmd.Process.Pid, syscall.SIGKILL)
+	})
+	err = cmd.Start()
+
 	// delete random blob
 	// assumption: the repo contains "p" blobs to delete, else the test will fail
 	err = bm.DeleteBlob("")
@@ -65,29 +62,18 @@ func TestSnapshotFix(t *testing.T) {
 		t.FailNow()
 	}
 
-	snapshot, _, err := bm.TakeSnapshot(bm.PathToTakeSnapshot)
-	require.NoError(t, err)
-
-	_, err = bm.DeleteSnapshot(snapshot)
-	require.NoError(t, err)
-
 	// Create a temporary dir to restore a snapshot
 	restoreDir := makeDir("restore-data-")
 	if restoreDir == "" {
 		t.FailNow()
 	}
 
+	stdout, err := bm.RunMaintenance()
+	require.NoError(t, err)
+
 	// try to restore a snapshot, this should error out
-	stdout, err := bm.RestoreGivenOrRandomSnapshot("", restoreDir)
+	stdout, err = bm.RestoreGivenOrRandomSnapshot("", restoreDir)
 	require.Error(t, err)
-
-	// stdout, err := bm.RunMaintenance()
-
-	// snapshot, _, err = bm.TakeSnapshot(bm.PathToTakeSnapshot)
-	// require.NoError(t, err)
-
-	// stdout, err := bm.RunMaintenance()
-	// require.Error(t, err)
 
 	// extract out object ID needed to be used in snapshot fix command
 	blobID := getBlobIDToBeDeleted(stdout)
@@ -112,7 +98,7 @@ func TestSnapshotFixInvalidFiles(t *testing.T) {
 		t.FailNow()
 	}
 
-	bm, err := blobmanipulator.NewBlobManipulator(baseDir)
+	bm, err := blobmanipulator.NewBlobManipulator(baseDir, dataRepoPath)
 	if err != nil {
 		if errors.Is(err, kopiarunner.ErrExeVariableNotSet) {
 			log.Println("Skipping recovery tests because KOPIA_EXE is not set")
@@ -120,16 +106,6 @@ func TestSnapshotFixInvalidFiles(t *testing.T) {
 			log.Println("Error creating Blob Manipulator:", err)
 		}
 
-		t.FailNow()
-	}
-
-	bm.DirCreater = getSnapshotter(baseDir, dataRepoPath)
-	if bm.DirCreater == nil {
-		t.FailNow()
-	}
-
-	err = bm.ConnectOrCreateRepo(dataRepoPath)
-	if err != nil {
 		t.FailNow()
 	}
 
@@ -141,6 +117,14 @@ func TestSnapshotFixInvalidFiles(t *testing.T) {
 	if err != nil {
 		t.FailNow()
 	}
+
+	kopiaExe := os.Getenv("KOPIA_EXE")
+	// cmd := exec.Command(kopiaExe, "snapshot", "create", bm.PathToTakeSnapshot)
+	cmd := exec.Command(kopiaExe, "maintenance", "run", "--full", "--force", "--safety", "none")
+	time.AfterFunc(10*time.Millisecond, func() {
+		syscall.Kill(cmd.Process.Pid, syscall.SIGKILL)
+	})
+	err = cmd.Start()
 
 	// delete random blob
 	// assumption: the repo contains "p" blobs to delete, else the test will fail
@@ -159,9 +143,6 @@ func TestSnapshotFixInvalidFiles(t *testing.T) {
 	// try to restore a snapshot, this should error out
 	stdout, err := bm.RestoreGivenOrRandomSnapshot("", restoreDir)
 	require.Error(t, err)
-
-	// stdout, err := bm.RunMaintenance()
-	// require.Error(t, err)
 
 	// fix all the invalid files
 	stdout, err = bm.SnapshotFixInvalidFiles("--verify-files-percent=100")
@@ -202,24 +183,4 @@ func getBlobIDToBeDeleted(stdout string) string {
 	}
 
 	return strings.TrimSpace(s1[wantedIndex])
-}
-
-func getSnapshotter(baseDirPath string, dataRepoPath string) *snapmeta.KopiaSnapshotter {
-	ks, err := snapmeta.NewSnapshotter(baseDirPath)
-	if err != nil {
-		if errors.Is(err, kopiarunner.ErrExeVariableNotSet) {
-			log.Println("Skipping recovery tests because KOPIA_EXE is not set")
-		} else {
-			log.Println("Error creating kopia Snapshotter:", err)
-		}
-
-		return nil
-	}
-
-	if err = ks.ConnectOrCreateRepo(dataRepoPath); err != nil {
-		log.Println("Error initializing kopia Snapshotter:", err)
-		return nil
-	}
-
-	return ks
 }
